backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
redis
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
